+++ combined-consent-generation.py
@@ -1,12 +1,14 @@
 import csv
 import json
-from model.helper import del_none
 from model.UiDataModel import TermCode
-from model.MappingDataModel import FhirMapping, FixedFHIRCriteria, CQLTimeRestrictionParameter
+from model.MappingDataModel import FhirMapping, FixedFHIRCriteria, CQLTimeRestrictionParameter, SimpleCardinality
 from model.MappingDataModel import CQLMapping, FixedCQLCriteria
 from model.TreeMap import TreeMap, TermEntryNode
 import argparse
 import os
+
+from util.codec.json import JSONFhirOntoEncoder
+
 
 def configure_args_parser():
     arg_parser = argparse.ArgumentParser(description='Generate the consent for the MII-FDPG')
@@ -51,7 +53,7 @@ def generate_fixed_fhir_criteria(provisions_code, provisions_display) -> list[Fi
 
 
 def generate_fixed_cql_criteria(provisions_code, provisions_display) -> list[FixedCQLCriteria]:
-    return [FixedCQLCriteria("Coding", "provision.provision.code.coding",
+    return [FixedCQLCriteria({"CodeableConcept"}, "provision.provision.code", SimpleCardinality.MANY,
             [{"code": code, "display": display, "system": "urn:oid:2.16.840.1.113883.3.1937.777.24.5.3"}])
             for code, display in zip(provisions_code, provisions_display)]
 
@@ -83,7 +85,7 @@ def process_csv(csv_file: str):
             fhir_mapping.timeRestrictionParameter = "date"
             cql_mapping.key = term_code
             cql_mapping.context = context
-            cql_mapping.timeRestriction = CQLTimeRestrictionParameter("Consent.datetime", ["dateTime"])
+            cql_mapping.timeRestriction = CQLTimeRestrictionParameter("datetime", {"dateTime"})
             cql_mapping.resourceType = "Consent"
             cql_mapping.primaryCode= {
                             "code": "57016-8",
@@ -114,7 +116,7 @@ def process_csv(csv_file: str):
 
 def save_json(filename: str, data):
     with open(filename, "w+", encoding='UTF-8') as f:
-        json.dump(data, f, default=lambda o: del_none(o.__dict__))
+        json.dump(data, f, cls=JSONFhirOntoEncoder)
 
 
 def append_to_json(filename: str, input_filename: str, data):